backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
sqlalchemy>=2.0.25
alembic>=1.13.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
httpx>=0.26.0
loguru>=0.7.2
python-dotenv>=1.0.1
google-generativeai>=0.3.2
psycopg2-binary>=2.9.9
aiofiles>=23.2.1
requests>=2.31.0
sentence-transformers>=2.2.2
numpy>=1.26.0
gunicorn>=21.2.0
email-validator>=2.1.0
